move editTransaction logic from LLC to coin-evm
