+++ apps/ledger-live-desktop/src/renderer/families/evm/EditTransaction/Body.tsx
@@ -1,3 +1,9 @@
+import {
+  getEditTransactionStatus,
+  hasMinimumFundsToCancel,
+  hasMinimumFundsToSpeedUp,
+  isTransactionConfirmed,
+} from "@ledgerhq/coin-evm/editTransaction/index";
 import { fromTransactionRaw } from "@ledgerhq/coin-evm/transaction";
 import { Transaction, TransactionRaw } from "@ledgerhq/coin-evm/types/index";
 import { UserRefusedOnDevice } from "@ledgerhq/errors";
@@ -5,12 +11,6 @@ import { getAccountCurrency } from "@ledgerhq/live-common/account/helpers";
 import { addPendingOperation, getMainAccount } from "@ledgerhq/live-common/account/index";
 import { SyncSkipUnderPriority } from "@ledgerhq/live-common/bridge/react/index";
 import useBridgeTransaction from "@ledgerhq/live-common/bridge/useBridgeTransaction";
-import {
-  getEditTransactionStatus,
-  hasMinimumFundsToCancel,
-  hasMinimumFundsToSpeedUp,
-  isTransactionConfirmed,
-} from "@ledgerhq/live-common/families/evm/editTransaction/index";
 import { Device } from "@ledgerhq/live-common/hw/actions/types";
 import { isNftTransaction } from "@ledgerhq/live-common/nft/index";
 import { isOldestPendingOperation } from "@ledgerhq/live-common/operation";
