+++ apps/ledger-live-desktop/src/renderer/families/evm/EditTransaction/steps/StepMethod.tsx
@@ -1,7 +1,7 @@
+import { getEditTransactionPatch } from "@ledgerhq/coin-evm/editTransaction/index";
 import { Transaction as EvmTransaction } from "@ledgerhq/coin-evm/types/index";
 import { getMainAccount } from "@ledgerhq/live-common/account/index";
 import { getAccountBridge } from "@ledgerhq/live-common/bridge/index";
-import { getEditTransactionPatch } from "@ledgerhq/live-common/families/evm/editTransaction/index";
 import { Flex } from "@ledgerhq/react-ui";
 import { AccountBridge } from "@ledgerhq/types-live";
 import invariant from "invariant";
