+++ apps/ledger-live-mobile/src/families/evm/EditTransactionFlow/MethodSelection.tsx
@@ -10,7 +10,7 @@ import {
   hasMinimumFundsToCancel,
   hasMinimumFundsToSpeedUp,
   isTransactionConfirmed,
-} from "@ledgerhq/live-common/families/evm/editTransaction/index";
+} from "@ledgerhq/coin-evm/editTransaction/index";
 import { fromTransactionRaw } from "@ledgerhq/live-common/transaction/index";
 import { getEnv } from "@ledgerhq/live-env";
 import { log } from "@ledgerhq/logs";
