+++ libs/coin-evm/src/editTransaction/index.ts
@@ -1,7 +1,7 @@
-import { getMinFees } from "@ledgerhq/coin-evm/getMinEditTransactionFees";
-import { getEditTransactionStatus } from "@ledgerhq/coin-evm/getTransactionStatus";
+import { getEditTransactionStatus } from "../getTransactionStatus";
 import { getEditTransactionPatch } from "./getEditTransactionPatch";
 import { getFormattedFeeFields } from "./getFormattedFeeFields";
+import { getMinFees } from "./getMinEditTransactionFees";
 import { hasMinimumFundsToCancel, hasMinimumFundsToSpeedUp } from "./hasMinimumFunds";
 import { isStrategyDisabled } from "./isStrategyDisabled";
 import { isTransactionConfirmed } from "./isTransactionConfirmed";
